Potential Gonotrophic Cycle Plus Trap Implementation
Implemented Target class with a trap agent as the child, in the future will have other mammals as children to the Target class too. I currently have a problem with deleting midges that are supposed to die from the model scheduler, and I cannot figure it out right now. I will keep working on this because none of the model can work without it.