Bug Fixes plus Switch to Continuous Space
I am currently dealing with bug issues. I fixed the problem where the midge was being killed twice, first by the trap and then by the death function. Afterwards another bug showed up where when the midge was killed by the trap the trap would then be deleted. I then switched over to a continuous space mode in an attempt to fix the bug but now the midge is being added to the kill list again. Will continue to try to fix it.