Bug Fixes Plus New Developments
I've fixed the issues with midges and traps, and now they are correctly deleted when they fly into the trap. I have also introduced Deer, a subclass of the Target class which are able to track the number of midge bites as well as whether they are carrying BTV. In addition, there is now an egg stage, which the midges lay in variable batch sizes and have survival rates of their own. I have added some comments to the code, but there is more to be added. In the future I plan on correctly simulating viremia, EIP of BTV, incorporating EHDV, and adding more big game animals. Next week we will probably focus on adding geographic features to the world and correctly scaling the simulation.
I have noticed that when the number of midges goes above 2000, there is a significant drop in performance. It is not currently a problem but in the future it would be wise to add Cython to speed up the numerically intensive tasks.